fix type

+++ Core/Src/dht11.c
@@ -59,5 +59,5 @@ DHT11 getDht11() {
     return (DHT11){DHT11_CHECKSUM_ERROR};
   }
 
-  return (DHT11){DHT11_OK, (int8_t)buffer[2], buffer[0]};
+  return (DHT11){DHT11_OK, buffer[2], buffer[0]};
 }
